restore: split tablesWorker into two
# New struct indexWorker is created out of original tablesWorker part
# responsible for restoring sstable files from single manifest index.
# Context of tablesWorker is to do the full restore where the context of indexWorker
# is to restore smaller chunk.
